+++ lib/Component/Payum/Stripe/Api.php
@@ -3,6 +3,7 @@
 use Payum\Core\Payum;
 use Payum\Core\Gateway;
 use Payum\Stripe\Request\Api\CreatePlan;
+use Payum\Stripe\Request\Api\CreateCustomer;
 use Stripe\Event as StripeEvent;
 use Http\Discovery\NotFoundException;
 
@@ -15,6 +16,7 @@ use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\GetCustomers;
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\GetPaymentMethods;
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\GetSubscriptions;
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\CancelSubscription;
+use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\GetConnectedAccounts;
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\GetWebhookEndpoints;
 use Vankosoft\PaymentBundle\Component\Payum\Stripe\Request\Api\CreateWebhookEndpoint;
 
@@ -144,6 +146,17 @@ final class Api
         return isset( $availableCustomers["data"] ) ? $availableCustomers["data"] : [];
     }
     
+    public function createCustomer( array $formData )
+    {
+        $custommer  = new \ArrayObject([
+            'name'  => $formData['name'],
+            'email' => $formData['email'],
+        ]);
+        $this->gateway->execute( $createCustommerRequest = new CreateCustomer( $custommer ) );
+        
+        return $createCustommerRequest->getFirstModel()->getArrayCopy();
+    }
+    
     public function getPaymentMethods( array $params = [] )
     {
         $stripeRequest              = new \ArrayObject( $params );
@@ -170,6 +183,16 @@ final class Api
             "id"    => $id,
         ]);
         $this->gateway->execute( new CancelSubscription( $subscription ) );
+    }
+    
+    public function getConnectedAccounts()
+    {
+        $stripeRequest  = new \ArrayObject( [] );
+        $this->gateway->execute( $getAccountsRequest = new GetConnectedAccounts( $stripeRequest ) );
+        
+        $availableAccounts = $getAccountsRequest->getFirstModel()->getArrayCopy();
+        
+        return isset( $availableAccounts["data"] ) ? $availableAccounts["data"] : [];
     }
     
     public function getWebhookEndpoints()